requirements.txt
streamlit>=1.20
faiss-cpu
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
transformers>=4.34.0,<4.35.0
tokenizers>=0.13.0
safetensors>=0.4.0
scikit-learn
pandas
numpy
python-dotenv
requests
